backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
pydantic-settings>=2.10,<3.0
celery[redis]>=5.5,<6.0
redis>=6.4,<7.0
sqlalchemy>=2.0,<3.0
alembic>=1.16,<2.0
pgvector>=0.4,<1.0
psycopg[binary]>=3.2,<4.0
litellm>=1.79,<2.0
opentelemetry-api>=1.36,<2.0
opentelemetry-sdk>=1.36,<2.0
pytest>=8.4,<9.0
httpx>=0.28,<1.0
PyJWT[crypto]>=2.10,<3.0
